requirements.txt
alembic==1.3.1
attrs==19.3.0
blinker==1.4
certbot==0.28.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==1.7.1
dnspython==1.16.0
dominate==2.4.0
eventlet==0.25.1
Flask==1.1.1
Flask-Assets==2.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
josepy==1.1.0
keyring==10.1
keyrings.alt==1.3
Mako==1.1.0
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
more-itertools==8.0.2
packaging==19.2
parsedatetime==2.1
pathlib2==2.3.5
pbr==1.10.0
pluggy==0.13.1
protobuf==3.11.2
py==1.8.0
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
PyJWT==1.7.1
pyOpenSSL==16.2.0
pyparsing==2.4.5
pyRFC3339==1.0
pyScss==1.3.5
pyserial==3.4
pytest==5.3.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2016.7
pyxdg==0.25
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.7.0
RPi.GPIO==0.6.5
rq==1.2.0
SecretStorage==2.3.1
six==1.13.0
SQLAlchemy==1.3.11
typing==3.7.4.1
urllib3==1.25.7
uWSGI==2.0.18
visitor==0.1.3
wcwidth==0.1.7
webassets==2.0
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
mysqlclient==1.4.6
